requirements.txt
# Core AI/LLM Framework
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0

# RAG & Embeddings
chromadb==1.2.1
sentence-transformers==3.4.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
# unstructured==0.18.15  # Temporarily disabled - requires Python <3.13
python-magic==0.4.27

# Task Queue & Background Processing
celery==5.4.0
redis==5.2.0

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Configuration & Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.7.0

# Additional AI utilities
tiktoken==0.9.0
